Created... Right before I went back to simple mode.
Instead of trying to be all fancy, just use them as they come. This
should be as precise as it gets. Except maybe when readings are far
apart. But then, when we have multiple accels between gyros, it's going
to be difficult to compute. I guess what I was about to decide is the
imprecise mode. Precise mode chops